requirements.txt
PyYAML==5.1
redis==2.10.6
coverage==4.5.1